+++ examples/11_fs-router/src/pages/index.tsx
@@ -4,4 +4,10 @@ const Home = () => (
   </div>
 );
 
+export const getConfig = () => {
+  return {
+    render: 'dynamic',
+  } as const;
+};
+
 export default Home;