+++ packages/waku/src/lib/config.ts
@@ -21,6 +21,7 @@ export async function resolveConfigDev(config: Config) {
     distDir: 'dist',
     pagesDir: 'pages',
     apiDir: 'api',
+    slicesDir: '_slices',
     privateDir: 'private',
     rscBase: 'RSC',
     middleware: DEFAULT_MIDDLEWARE,