+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.25.0] - 2025-08-12
+
+### Added
+
+- Breaking: page defaults to static
+- Slice API in waku/router
+
 ## [0.24.0] - 2025-08-05
 
 ### Added